moved account page into (app) route group
